pyproject.toml
[project]
name = "dreamflow"
version = "0.1.0"
description = "Airflow Based Scheduler/Workflow/Task Orchestrater"
authors = [
    {name = "eric rodefeld",email = "eric.d.rodefeld@gmail.com"},
]
dependencies = [
    "apache-airflow[postgres,celery,redis,mongo]==2.10.4",
    "shapely==2.0.6",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.setuptools]
py-modules = []
